Speedbrake: B737 NG changed to use Spoiler L&R 4
While it is clear from the docstring that Spoiler L&R 3 were selected
for B737-NG due to L&R 4 having a short field option, Andrew has
provided a document reference which states that the opposite is True.
This is substantiated by the 737-7 frame where Spoiler L&R 3 exhibit the
short field behaviour. Data for all affected aircraft will be processed
on test for verification.