kpv: remove speedbrake deployed with conf duration.
Following on from 38ece73, it was unclear whether 2.0 was originally
referring to CONF 2 or the raw values for the configuration mapping
which would make it >= CONF 0. For this to be robust, the raw value
should have been determined by lookup in the mapping using the
appropriate state name, e.g.

    at.constants.AVAILABLE_CONF_STATES_REV['2']

We are not using this for any event, so just remove it.

Make use of 'Speedbrake Deployed With Flap Duration' instead.