split: add validation_dt separate from fallback_dt
We need a date and time for validating the time parameters, e.g.
ensuring that the year in the data is not in the future. Typically we
want to use the date and time that the data was uploaded into the
system. Prior to this we were using the fallback date and time, but this
value could be obtained from the filename and may be equally unreliable.

Using the same value for validation as well as a fallback could lead to
incorrectly invalidating time parameters resulting in erroneous timebase
calculations.