split: add assert as follow up to ffb79b5
The fallback date and time cannot be ahead of the date and time that we
are using for validation of the date and time parameters in the data.
Typically the validation date and time should be when the raw data file
was uploaded to the system for processing, thus the fallback cannot be
ahead of this.

Also tweak the assertion for the future check to be <= instead of <.