Temporary fix to fallback in future during splitting
Padding to superframe boundaries during splitting was not being
taken into account when calculating each splits adjusted fallback
datetime. This drift in fallback datetime calculation resulted in
datetimes in the future.

Adjustments of fallback datetimes in relation to each split will be
made using the duration of the split slice(without padding) and the
padding introduced at the start of the new segment. This can result in
small overlaps in time between segments (<128secs) but this is
reflective of the actual flight data in each segment.

Although not pretty this commit will resolve the problem until a
comprehensive refactor of splitting can be implemented.