Changed loc start to loc estab when looking for 2 dots runway turnoff
due to approach and landing slice finishing at 45 degree turning
off runway the first time we may see 2 dots is before establised
but after loc start.

In the case where we are still established at the end of the approach
and landing phase we will use the last valid sample within this slice.