splitter: fix fallback if date and time parameters constant.
We should only allow constant date and time parameters to affect the
fallback for certain recorders that are "frame doubled" to provide an
alternative date and time where there is none in the data.

Constant date and time parameters in the data should be considered bad
and the fallback used in preference.